fix: use float values to avoid lossing presicion on operation

+++ aggregation_mode/proof_aggregator/src/backend/mod.rs
@@ -551,13 +551,13 @@ impl ProofAggregator {
             .await
             .map_err(|e| AggregatedProofSubmissionError::GasPriceError(e.to_string()))?;
 
-        let new_base_fee = current_gas_price * (1 + base_bump_percentage as u128 / 100);
-        let new_max_fee = new_base_fee * (1 + max_fee_bump_percentage as u128 / 100);
+        let new_base_fee = current_gas_price as f64 * (1.0 + base_bump_percentage as f64 / 100.0);
+        let new_max_fee = new_base_fee as f64 * (1.0 + max_fee_bump_percentage as f64 / 100.0);
         let new_priority_fee = priority_fee_wei;
 
         Ok(tx_req
-            .with_max_fee_per_gas(new_max_fee)
-            .with_max_priority_fee_per_gas(new_priority_fee))
+            .with_max_fee_per_gas(new_max_fee as u128)
+            .with_max_priority_fee_per_gas(new_priority_fee as u128))
     }
 
     async fn wait_until_can_submit_aggregated_proof(
